Refactor skim package seams
