+++ tests/test_trajectory_viewer.py
@@ -1,287 +1,28 @@
-"""Tests for skim."""
+"""Trajectory-viewer tests for skim.
+
+This module covers trajectory normalization, tree construction, detail rendering,
+local viewer controls, and specialized drag/keyboard behavior. It does not own
+outer app-shell layout or generic preview classification.
+"""
 
 import json
-from typing import Any
 
+from conftest import (
+    _all_collapsible_titles,
+    _collapsible_by_title,
+    _detail_syntax_blocks,
+    _detail_text,
+    _static_content,
+    _top_level_collapsible_titles,
+    _tree_labels,
+    multi_step_trajectory,
+    sample_trajectory,
+)
 from rich.syntax import Syntax
 from rich.text import Text
-from textual.widgets import Collapsible, Markdown, Static, Tree
+from textual.widgets import Markdown, Static, Tree
 
-from skim import (
-    PreviewPane,
-    SkimApp,
-    SubmissionSummary,
-    TrajectoryViewer,
-    normalize_events,
-    render_file,
-)
-
-
-async def test_app_launches():
-    """App starts without crashing."""
-    app = SkimApp(path=".")
-    async with app.run_test():
-        assert app.title == "skim"
-        assert app.grid == [["pane-0"]]
-
-
-async def test_split_right():
-    """Pressing s then right creates a second pane."""
-    app = SkimApp(path=".")
-    async with app.run_test() as pilot:
-        await pilot.press("s")
-        await pilot.press("right")
-        assert app._total_panes() == 2
-        assert len(app.grid[0]) == 2
-
-
-async def test_split_down():
-    """Pressing s then down creates a second row."""
-    app = SkimApp(path=".")
-    async with app.run_test() as pilot:
-        await pilot.press("s")
-        await pilot.press("down")
-        assert app._total_panes() == 2
-        assert len(app.grid) == 2
-
-
-async def test_close_pane():
-    """Can close a pane but not the last one."""
-    app = SkimApp(path=".")
-    async with app.run_test() as pilot:
-        # open a split
-        await pilot.press("s")
-        await pilot.press("right")
-        assert app._total_panes() == 2
-
-        # close it
-        await pilot.press("d")
-        assert app._total_panes() == 1
-
-        # can't close the last one
-        await pilot.press("d")
-        assert app._total_panes() == 1
-
-
-async def test_cycle_pane():
-    """w cycles through panes."""
-    app = SkimApp(path=".")
-    async with app.run_test() as pilot:
-        await pilot.press("s")
-        await pilot.press("right")
-        first_active = app.active_pane_id
-
-        await pilot.press("w")
-        second_active = app.active_pane_id
-
-        assert first_active != second_active
-
-        await pilot.press("w")
-        assert app.active_pane_id == first_active
-
-
-async def test_max_panes():
-    """Cannot exceed 6 panes."""
-    app = SkimApp(path=".")
-    async with app.run_test() as pilot:
-        for _ in range(5):
-            await pilot.press("s")
-            await pilot.press("right")
-        assert app._total_panes() == 6
-
-        # seventh should fail
-        await pilot.press("s")
-        await pilot.press("right")
-        assert app._total_panes() == 6
-
-
-async def test_file_preview(tmp_path):
-    """Selecting a file shows its content."""
-    test_file = tmp_path / "hello.txt"
-    test_file.write_text("hello world")
-
-    app = SkimApp(path=str(tmp_path))
-    async with app.run_test() as pilot:
-        # give the tree time to load
-        await pilot.pause()
-        from skim import PreviewPane
-
-        pane = app.query_one(f"#{app.active_pane_id}", PreviewPane)
-        pane.show_file(test_file)
-        assert pane.current_path == test_file
-
-
-async def test_f_enters_file_tree_mode(tmp_path):
-    """Pressing f should focus the file tree and enter tree mode."""
-    (tmp_path / "one.txt").write_text("x")
-    app = SkimApp(path=str(tmp_path))
-
-    async with app.run_test() as pilot:
-        await pilot.pause()
-        pane = app.query_one(f"#{app.active_pane_id}", PreviewPane)
-        tree = app.query_one("DirectoryTree")
-
-        assert app.focused is pane
-        assert not app.file_tree_mode
-
-        await pilot.press("f")
-        await pilot.pause()
-
-        assert app.file_tree_mode
-        assert app.focused is tree
-
-
-async def test_file_tree_mode_up_down_moves_tree_cursor(tmp_path):
-    """While file-tree mode is active, up/down should move the tree cursor."""
-    for index in range(4):
-        (tmp_path / f"file_{index}.txt").write_text("x")
-    app = SkimApp(path=str(tmp_path))
-
-    async with app.run_test() as pilot:
-        await pilot.pause()
-        tree = app.query_one("DirectoryTree")
-
-        await pilot.press("f")
-        await pilot.pause()
-        before = tree.cursor_line
-        await pilot.press("down")
-        await pilot.pause()
-
-        assert tree.cursor_line > before
-
-
-async def test_file_tree_mode_enter_opens_file_and_returns_to_pane(tmp_path):
-    """Entering a file from tree mode should open it and restore pane focus."""
-    test_file = tmp_path / "open-me.txt"
-    test_file.write_text("hello")
-    app = SkimApp(path=str(tmp_path))
-
-    async with app.run_test() as pilot:
-        await pilot.pause()
-        pane = app.query_one(f"#{app.active_pane_id}", PreviewPane)
-
-        await pilot.press("f")
-        await pilot.pause()
-        await pilot.press("down")
-        await pilot.pause()
-        await pilot.press("enter")
-        await pilot.pause()
-
-        assert pane.current_path == test_file
-        assert not app.file_tree_mode
-        assert app.focused is pane
-
-
-async def test_file_tree_mode_escape_returns_to_active_pane(tmp_path):
-    """Escape should leave file-tree mode and return focus to the preview pane."""
-    (tmp_path / "one.txt").write_text("x")
-    app = SkimApp(path=str(tmp_path))
-
-    async with app.run_test() as pilot:
-        await pilot.pause()
-        pane = app.query_one(f"#{app.active_pane_id}", PreviewPane)
-        footer = app.query_one("#status-bar", Static)
-
-        await pilot.press("f")
-        await pilot.pause()
-        assert app.file_tree_mode
-        assert "Back" in _static_content(footer)
-
-        await pilot.press("escape")
-        await pilot.pause()
-
-        assert not app.file_tree_mode
-        assert app.focused is pane
-        assert "File tree" in _static_content(footer)
-
-
-async def test_down_outside_file_tree_mode_still_scrolls_preview_pane(tmp_path):
-    """Outside file-tree mode, down should keep scrolling the active pane."""
-    test_file = tmp_path / "long.txt"
-    test_file.write_text("\n".join(f"line {index}" for index in range(400)))
-    (tmp_path / "other.txt").write_text("x")
-    app = SkimApp(path=str(tmp_path))
-
-    async with app.run_test() as pilot:
-        await pilot.pause()
-        pane = app.query_one(f"#{app.active_pane_id}", PreviewPane)
-        tree = app.query_one("DirectoryTree")
-        pane.show_file(test_file)
-        await pilot.pause()
-
-        before_scroll = pane.scroll_y
-        before_cursor = tree.cursor_line
-        await pilot.press("down")
-        await pilot.pause()
-
-        assert pane.scroll_y > before_scroll
-        assert tree.cursor_line == before_cursor
-
-
-async def test_shift_down_still_moves_file_tree_cursor(tmp_path):
-    """Shift+down should remain a convenience shortcut for tree navigation."""
-    for index in range(4):
-        (tmp_path / f"file_{index}.txt").write_text("x")
-    app = SkimApp(path=str(tmp_path))
-
-    async with app.run_test() as pilot:
-        await pilot.pause()
-        tree = app.query_one("DirectoryTree")
-        before = tree.cursor_line
-
-        await pilot.press("shift+down")
-        await pilot.pause()
-
-        assert tree.cursor_line > before
-
-
-def test_generic_json_uses_syntax_preview(tmp_path):
-    """Generic JSON keeps the syntax-highlighted preview."""
-    test_file = tmp_path / "plain.json"
-    test_file.write_text(json.dumps({"hello": "world"}))
-
-    widgets = render_file(test_file)
-
-    assert len(widgets) == 1
-    assert isinstance(widgets[0], Static)
-    assert isinstance(_static_content(widgets[0]), Syntax)
-
-
-def test_invalid_json_falls_back_to_syntax_preview(tmp_path):
-    """Invalid JSON still renders through the generic JSON path."""
-    test_file = tmp_path / "broken.json"
-    test_file.write_text("{not json")
-
-    widgets = render_file(test_file)
-
-    assert len(widgets) == 1
-    assert isinstance(widgets[0], Static)
-    assert isinstance(_static_content(widgets[0]), Syntax)
-
-
-def test_wrapped_trajectory_json_uses_trajectory_viewer(tmp_path):
-    """A wrapped raw trajectory uses the trajectory viewer."""
-    test_file = tmp_path / "output.json"
-    test_file.write_text(json.dumps({"trajectory": sample_trajectory()}))
-
-    widgets = render_file(test_file)
-
-    assert len(widgets) == 1
-    assert isinstance(widgets[0], TrajectoryViewer)
-    assert len(widgets[0].events) == 4
-    assert _tree_labels(widgets[0]._tree) == ["Metadata", "Final Output", "Step 1"]
-
-
-def test_bare_trajectory_json_uses_trajectory_viewer(tmp_path):
-    """A bare trajectory object uses the trajectory viewer."""
-    test_file = tmp_path / "trajectory.json"
-    test_file.write_text(json.dumps(sample_trajectory()))
-
-    widgets = render_file(test_file)
-
-    assert len(widgets) == 1
-    assert isinstance(widgets[0], TrajectoryViewer)
-    assert len(widgets[0].events) == 4
+from skim import PreviewPane, SkimApp, TrajectoryViewer, normalize_events
 
 
 def test_normalize_events_extracts_supported_event_kinds():
@@ -601,65 +342,6 @@ async def test_unmatched_function_call_result_still_renders_safely():
     assert [child.label.plain for child in step.children[2].children] == ["Input", "Output"]
 
 
-def test_submission_json_uses_submission_summary(tmp_path):
-    """Worker submission JSON uses the submission summary viewer."""
-    test_file = tmp_path / "submission.json"
-    test_file.write_text(
-        json.dumps(
-            {
-                "task_name": "Factors affecting WHPs",
-                "submission_type": "Complete task",
-                "prompt": "Compare spray diary information.",
-                "quick_scores": "Model A: 32%",
-                "agentic_grader_guidance": "Identify Chlorpyrifos.",
-                "load_trajectories_s3": "https://example.invalid/trajectory.json",
-            }
-        )
-    )
-
-    widgets = render_file(test_file)
-
-    assert len(widgets) == 1
-    assert isinstance(widgets[0], SubmissionSummary)
-    text = widgets[0].summary_text.plain
-    assert "Factors affecting WHPs" in text
-    assert "https://example.invalid/trajectory.json" in text
-
-
-async def test_split_panes_keep_specialized_preview(tmp_path):
-    """Specialized previews still work inside split panes."""
-    test_file = tmp_path / "output.json"
-    test_file.write_text(json.dumps({"trajectory": sample_trajectory()}))
-    app = SkimApp(path=str(tmp_path))
-
-    async with app.run_test() as pilot:
-        await pilot.press("s")
-        await pilot.press("right")
-        pane = app.query_one(f"#{app.active_pane_id}", PreviewPane)
-        pane.show_file(test_file)
-
-        assert app._total_panes() == 2
-        assert isinstance(pane.query_one(TrajectoryViewer), TrajectoryViewer)
-
-
-async def test_scroll_keys_scroll_generic_preview_pane(tmp_path):
-    """App scroll actions should still scroll a normal preview pane."""
-    test_file = tmp_path / "long.txt"
-    test_file.write_text("\n".join(f"line {index}" for index in range(400)))
-    app = SkimApp(path=str(tmp_path))
-
-    async with app.run_test() as pilot:
-        pane = app.query_one(f"#{app.active_pane_id}", PreviewPane)
-        pane.show_file(test_file)
-        await pilot.pause()
-
-        before = pane.scroll_y
-        await pilot.press("down")
-        await pilot.pause()
-
-        assert pane.scroll_y > before
-
-
 async def test_scroll_keys_scroll_trajectory_detail_panel():
     """App scroll actions should scroll the trajectory detail panel, not only the outer pane."""
     stdout = "\n".join(f"line {index}" for index in range(200))
@@ -789,85 +471,6 @@ async def test_trajectory_footer_switches_between_tree_and_detail_modes():
         assert "Branch" in reset_content.plain
 
 
-async def test_global_footer_only_shows_app_wide_commands():
-    """Global footer should not include trajectory-specific commands."""
-    app = SkimApp(path=".")
-
-    async with app.run_test() as pilot:
-        await pilot.pause()
-
-        footer = app.query_one("#status-bar", Static)
-        content = _static_content(footer)
-
-        assert isinstance(content, str)
-        assert "Scroll" in content
-        assert "File tree" in content
-        assert "Open" in content
-        assert "JSON" not in content
-        assert "Branch" not in content
-        assert "Detail" not in content
-        assert "Esc" not in content
-
-
-async def test_non_trajectory_previews_do_not_mount_local_footer(tmp_path):
-    """Only trajectory previews should render a local footer."""
-    generic = tmp_path / "data.json"
-    generic.write_text('{"alpha": 1}')
-    submission = tmp_path / "submission.json"
-    submission.write_text('{"task_name": "Task", "submission_type": "worker", "prompt": "Prompt"}')
-    app = SkimApp(path=str(tmp_path))
-
-    async with app.run_test() as pilot:
-        pane = app.query_one(f"#{app.active_pane_id}", PreviewPane)
-
-        pane.show_file(generic)
-        await pilot.pause()
-        assert not pane.query(".trajectory-footer")
-
-        pane.show_file(submission)
-        await pilot.pause()
-        assert not pane.query(".trajectory-footer")
-
-
-async def test_mouse_drag_scrolls_preview_pane(tmp_path):
-    """Dragging inside a preview pane should scroll long generic content."""
-    test_file = tmp_path / "long.txt"
-    test_file.write_text("\n".join(f"line {index}" for index in range(400)))
-    app = SkimApp(path=str(tmp_path))
-
-    async with app.run_test() as pilot:
-        pane = app.query_one(f"#{app.active_pane_id}", PreviewPane)
-        pane.show_file(test_file)
-        await pilot.pause()
-        before = pane.scroll_y
-
-        await pilot.mouse_down(pane, offset=(5, 10))
-        await pilot.hover(pane, offset=(5, 1))
-        await pilot.mouse_up(pane, offset=(5, 1))
-        await pilot.pause()
-
-        assert pane.scroll_y > before
-
-
-async def test_mouse_drag_scrolls_directory_tree(tmp_path):
-    """Dragging inside the file tree should scroll the tree."""
-    for index in range(80):
-        (tmp_path / f"file_{index}.txt").write_text("x")
-    app = SkimApp(path=str(tmp_path))
-
-    async with app.run_test() as pilot:
-        await pilot.pause()
-        tree = app.query_one("DirectoryTree")
-        before = tree.scroll_y
-
-        await pilot.mouse_down(tree, offset=(5, 10))
-        await pilot.hover(tree, offset=(5, 1))
-        await pilot.mouse_up(tree, offset=(5, 1))
-        await pilot.pause()
-
-        assert tree.scroll_y > before
-
-
 async def test_mouse_drag_scrolls_trajectory_tree():
     """Dragging inside the trajectory tree should scroll long step lists."""
     viewer = TrajectoryViewer(multi_step_trajectory(20))
@@ -909,123 +512,3 @@ async def test_mouse_drag_scrolls_trajectory_detail():
         await pilot.pause()
 
         assert detail.scroll_y > before
-
-
-def sample_trajectory(
-    stdout: str = "plain terminal output",
-    arguments: dict | None = None,
-    output: dict | None = None,
-    include_call: bool = True,
-    include_result: bool = True,
-):
-    """Return a small raw trajectory fixture."""
-    arguments = {"command": "ls -la"} if arguments is None else arguments
-    output = (
-        {"stdout": stdout, "stderr": "warning text", "returncode": 0} if output is None else output
-    )
-    step_output = [
-        {
-            "type": "reasoning",
-            "content": [
-                {
-                    "type": "input_text",
-                    "text": "Need inspect files.\n\n- Check data",
-                }
-            ],
-        },
-        {
-            "type": "message",
-            "role": "assistant",
-            "content": [
-                {
-                    "type": "output_text",
-                    "text": "I will inspect files.\n\n```bash\nls -la\n```",
-                }
-            ],
-        },
-    ]
-    if include_call:
-        step_output.append(
-            {
-                "type": "function_call",
-                "name": "syntara__executeBash",
-                "callId": "tool-1",
-                "status": "completed",
-                "arguments": json.dumps(arguments),
-            }
-        )
-    if include_result:
-        step_output.append(
-            {
-                "type": "function_call_result",
-                "name": "syntara__executeBash",
-                "callId": "tool-1",
-                "status": "completed",
-                "output": {"text": json.dumps(output)},
-            }
-        )
-    return {
-        "metadata": {
-            "trajectory_id": "traj-1",
-            "task_id": "task-1",
-            "llm_provider": "anthropic",
-            "llm_model": "claude-test",
-            "input_tokens": 100,
-            "output_tokens": 200,
-        },
-        "context_compaction_events": [{"message_index": 3}],
-        "final_output": "## Final\n\nDone.",
-        "steps": [{"output": step_output}],
-    }
-
-
-def multi_step_trajectory(step_count: int) -> dict[str, Any]:
-    """Return a trajectory with repeated steps to force tree overflow."""
-    trajectory = sample_trajectory()
-    step = trajectory["steps"][0]
-    trajectory["steps"] = [step for _ in range(step_count)]
-    return trajectory
-
-
-def _static_content(widget: Static):
-    return widget._Static__content
-
-
-def _tree_labels(tree: Tree):
-    return [child.label.plain for child in tree.root.children]
-
-
-def _detail_text(viewer: TrajectoryViewer) -> str:
-    parts = []
-    for widget in viewer._detail_wrap.query(Static):
-        content = _static_content(widget)
-        if isinstance(content, Text):
-            parts.append(content.plain)
-        elif isinstance(content, Syntax):
-            parts.append(content.code)
-        else:
-            parts.append(str(content))
-    return "\n".join(parts)
-
-
-def _detail_syntax_blocks(viewer: TrajectoryViewer) -> list[Syntax]:
-    return [
-        _static_content(widget)
-        for widget in viewer._detail_wrap.query(Static)
-        if isinstance(_static_content(widget), Syntax)
-    ]
-
-
-def _top_level_collapsible_titles(viewer: TrajectoryViewer) -> list[str]:
-    return [child.title for child in viewer._detail_wrap.children if isinstance(child, Collapsible)]
-
-
-def _all_collapsible_titles(viewer: TrajectoryViewer) -> list[str]:
-    return [widget.title for widget in viewer._detail_wrap.query(Collapsible)]
-
-
-def _collapsible_by_title(viewer: TrajectoryViewer, title: str) -> Collapsible:
-    for widget in viewer._detail_wrap.query(Collapsible):
-        if widget.title == title:
-            return widget
-    raise AssertionError(f"Collapsible with title {title!r} not found")
